// app/navigation/types.ts
export enum screens {
  Init = 'Init',
  Signup = 'SignupOptions',
  Login = 'LoginOptions',
  CreateAccount = 'signup',
  Workouts = 'Workouts',
  WhatchWorkouts = 'WatchWorkout',
  WhatchExercise = 'Watch_Exercise',
  WhatchWorkoutSummary = 'Watch_Summary',
  DoWorkout = 'Do_Workout',
  Account = 'Account',
  CompleteAccount = 'Complete_Account',
  Settings = 'Settings',
  Plans = 'Programs',
  WhatchPlan = 'Watch_Plan',
  WhatchPlanAi = 'Watch_Plan_AI',
  Coach = 'Coach',
  CoachChat = 'Chat',
  Feed = 'Home',
  WhatchRecipe = 'FaceRecipe',
  WhatchArticle = 'Article',
  Exercises = 'Exercises',
  StoreModal = 'Store_Modal',
}